+++ CHANGELOG.md
@@ -14,7 +14,7 @@
 
 - Fixed issue when last social icon was above others in the list, #453
 - Fixed issue when 'or as guest' was showing when guest/wordpress only option was selected 
-
+- Other small fixes & improvements
 
 ## 0.1.27 – 18.05.2020