+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 
 **Fixes:**
 
+- Now using dark hardcoded color for text inside comment editor to ensure it looks good for dark themes, #457
 - Fixed issue when last social icon was above others in the list, #453
 - Fixed issue when 'or as guest' was showing when guest/wordpress only option was selected 
 - Other small fixes & improvements